+++ python/src/nanoarrow/c_array.py
@@ -31,7 +31,7 @@ from nanoarrow._lib import (
     _obj_is_capsule,
 )
 from nanoarrow.c_buffer import c_buffer
-from nanoarrow.c_lib import c_schema, c_schema_view
+from nanoarrow.c_schema import c_schema, c_schema_view
 
 
 def c_array(obj, schema=None) -> CArray:
