+++ python/src/nanoarrow/c_schema.py
