+++ python/src/nanoarrow/device.py
@@ -17,7 +17,7 @@
 
 from nanoarrow._lib import DEVICE_CPU, CDeviceArray, Device, DeviceType  # noqa: F401
 from nanoarrow.c_array import c_array
-from nanoarrow.c_lib import c_schema
+from nanoarrow.c_schema import c_schema
 
 
 def cpu():
